a_small_project
Jokenpô is a game in which people play with their hands, choosing between rock, 
paper and scissors. And it works like this: the scissors cut the paper, 
but break with the stone; the paper wraps the rock 
but is cut by the scissors and the rock breaks the scissors and is wrapped by the paper.